Extract fixtures
